+++ backtesting/_stats.py
@@ -15,7 +15,7 @@ def compute_drawdown_duration_peaks(dd: pd.Series):
     iloc = np.unique(np.r_[(dd == 0).values.nonzero()[0], len(dd) - 1])
     iloc = pd.Series(iloc, index=dd.index[iloc])
     df = iloc.to_frame('iloc').assign(prev=iloc.shift())
-    df = df[df['iloc'] > df['prev'] + 1].astype(int)
+    df = df[df['iloc'] > df['prev'] + 1].astype(np.int64)
 
     # If no drawdown since no trade, avoid below for pandas sake and return nan series
     if not len(df):